Netomnia-Testing-Main/requirements.txt
selenium==4.16.0
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.1.0
certifi==2023.11.17
idna==3.6
charset-normalizer==3.3.2
webdriver-manager==4.0.1
ImageHash==4.3.1
pdf2image==1.17.0
PyMuPDF==1.24.9
Pillow==10.1.0
python-pptx==0.6.23
pandas==2.3.3
PyPDF2==3.0.1